requirements.txt
# Core
pandas>=2.2
pyarrow>=15.0
polars>=1.6.0
numpy>=1.26
pydantic>=2.7
pydantic-settings>=2.2
python-dotenv>=1.0
pyyaml>=6.0.1

# API
fastapi>=0.111
uvicorn[standard]>=0.30
scikit-learn>=1.5
lightgbm>=4.5

# MLOps
mlflow>=2.14

# Data quality (elige uno; dejamos ambos y decides)
great-expectations>=0.18
pandera>=0.17

# Utilidades
haversine>=2.8.1
joblib>=1.4
